+++ ddht/v5/message_dispatcher.py
@@ -20,11 +20,12 @@ from trio.abc import ReceiveChannel, SendChannel
 from trio.hazmat import checkpoint
 
 from ddht.abc import NodeDBAPI
+from ddht.base_message import IncomingMessage, OutgoingMessage
 from ddht.constants import IP_V4_ADDRESS_ENR_KEY, UDP_PORT_ENR_KEY
+from ddht.endpoint import Endpoint
 from ddht.exceptions import UnexpectedMessage
 from ddht.typing import NodeID
 from ddht.v5.abc import ChannelHandlerSubscriptionAPI, MessageDispatcherAPI
-from ddht.v5.channel_services import Endpoint, IncomingMessage, OutgoingMessage
 from ddht.v5.constants import (
     MAX_NODES_MESSAGE_TOTAL,
     MAX_REQUEST_ID,
