+++ tox.ini
@@ -17,7 +17,7 @@ line_length=88
 [flake8]
 max-line-length= 100
 exclude= venv*,.tox,docs,build
-ignore=W503
+ignore=W503,E203
 
 [testenv]
 usedevelop=True
